+++ src/main/kotlin/space/kscience/gradle/KScienceProjectPlugin.kt
@@ -5,7 +5,6 @@ import kotlinx.validation.BinaryCompatibilityValidatorPlugin
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.api.publish.maven.MavenPom
-import org.gradle.api.publish.maven.tasks.PublishToMavenRepository
 import org.gradle.kotlin.dsl.*
 import org.jetbrains.changelog.ChangelogPlugin
 import org.jetbrains.changelog.ChangelogPluginExtension
@@ -18,7 +17,6 @@ import space.kscience.gradle.internal.*
  */
 public class KSciencePublishingExtension(public val project: Project) {
     private var isVcsInitialized = false
-    internal val repositoryNames = mutableSetOf<String>()
 
     /**
      * Configures Git repository (sources) for the publication.
@@ -55,18 +53,16 @@ public class KSciencePublishingExtension(public val project: Project) {
      *
      * @param githubProject the GitHub project.
      * @param githubOrg the GitHub user or organization.
-     * @param addToRelease publish packages in the `release` task to the GitHub repository.
+     * @param deploy publish packages in the `deploy` task to the GitHub repository.
      */
     public fun github(
         githubOrg: String,
         githubProject: String,
-        addToRelease: Boolean = project.requestPropertyOrNull("publishing.github") == "true",
+        deploy: Boolean = project.requestPropertyOrNull("publishing.github") == "true",
     ) {
-
-        if (addToRelease) {
+        if (deploy) {
             try {
                 project.addGithubPublishing(githubOrg, githubProject)
-                repositoryNames += "github"
             } catch (t: Throwable) {
                 project.logger.error("Failed to set up github publication", t)
             }
@@ -77,15 +73,12 @@ public class KSciencePublishingExtension(public val project: Project) {
      * Adds Space Packages Maven repository to publishing.
      *
      * @param spaceRepo the repository URL.
-     * @param addToRelease publish packages in the `release` task to the Space repository.
+     * @param deploy publish packages in the `deploy` task to the Space repository.
      */
     public fun space(
         spaceRepo: String,
-        addToRelease: Boolean = project.requestPropertyOrNull("publishing.space") != "false",
     ) {
         project.addSpacePublishing(spaceRepo)
-
-        if (addToRelease) repositoryNames += "space"
     }
 
     /**
@@ -93,13 +86,9 @@ public class KSciencePublishingExtension(public val project: Project) {
      *
      * @param addToRelease  publish packages in the `release` task to the Sonatype repository.
      */
-    public fun sonatype(
-        addToRelease: Boolean = (project.requestPropertyOrNull("publishing.sonatype") != "false"),
-    ) {
+    public fun sonatype() {
         require(isVcsInitialized) { "The project vcs is not set up use 'pom' method to do so" }
         project.addSonatypePublishing()
-
-        if (addToRelease) repositoryNames += "sonatype"
     }
 }
 
@@ -227,13 +216,6 @@ public open class KScienceProjectPlugin : Plugin<Project> {
             dependsOn(generateReadme)
         }
 
-
-        val releaseAll by tasks.creating {
-            group = RELEASE_GROUP
-            description = "Publish development or production release based on version suffix"
-            dependsOn(generateReadme)
-        }
-
         tasks.create("version") {
             group = "publishing"
             val versionFile = project.buildDir.resolve("project-version.txt")
@@ -242,34 +224,6 @@ public open class KScienceProjectPlugin : Plugin<Project> {
                 versionFile.createNewFile()
                 versionFile.writeText(project.version.toString())
                 println(project.version)
-            }
-        }
-
-        allprojects {
-            afterEvaluate {
-                ksciencePublish.repositoryNames.forEach { repositoryName ->
-                    val repositoryNameCapitalized = repositoryName.capitalize()
-
-                    val pattern = "publish(?<publication>.*)PublicationTo${repositoryNameCapitalized}Repository"
-                        .toRegex()
-
-                    tasks.withType<PublishToMavenRepository>().forEach forEachPublication@{
-                        val matchResult = pattern.matchEntire(it.name) ?: return@forEachPublication
-                        val publicationName = matchResult.groups["publication"]!!.value.capitalize()
-                        val releaseTaskName = "release$publicationName"
-
-                        val targetReleaseTask =
-                            rootProject.tasks.findByName(releaseTaskName) ?: rootProject.tasks.create(releaseTaskName) {
-                                group = RELEASE_GROUP
-                                description =
-                                    "Publish platform release artifact for $publicationName to all repositories"
-                            }
-
-                        releaseAll.dependsOn(targetReleaseTask)
-
-                        targetReleaseTask.dependsOn(it)
-                    }
-                }
             }
         }
 
@@ -288,6 +242,6 @@ public open class KScienceProjectPlugin : Plugin<Project> {
     }
 
     public companion object {
-        public const val RELEASE_GROUP: String = "release"
+        public const val DEPLOY_GROUP: String = "deploy"
     }
 }